+++ packages/storybook/CHANGELOG.md
@@ -4,6 +4,42 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+v0.52.6
+------------------------------
+*June 17, 2022*
+### Changed
+- updated vue.config to include all atoms in components using @use and @forward
+
+
+v0.52.5
+------------------------------
+*June 17, 2022*
+### Changed
+- updated vue.config to include f-image-tile in components using @use and @forward
+
+
+v0.52.4
+------------------------------
+*June 17, 2022*
+### Changed
+- updated vue.config to include f-form-field in components using @use and @forward
+
+
+v0.52.3
+------------------------------
+*June 17, 2022*
+### Changed
+- updated vue.config to include f-filter-pill in components using @use and @forward
+
+
+v0.52.2
+------------------------------
+*June 16, 2022*
+
+### Changed
+- updated vue.config to include f-error-message in components using @use and @forward
+
+
 v0.52.1
 ------------------------------
 *June 16, 2022*